requirements.txt
django
djangorestframework
psycopg2

djangorestframework-simplejwt